requirements.txt
# Velocity Engine Requirements
# Batch 1 + Batch 2

# === Core Dependencies ===
python-dotenv>=1.0.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# === Data & Indicators ===
pandas>=2.0.0
pandas-ta>=0.3.14b
numpy>=1.24.0

# === Broker APIs ===
alpaca-py>=0.21.0
yfinance>=0.2.36

# === Database ===
sqlalchemy>=2.0.0
psycopg2-binary>=2.9.9
alembic>=1.13.0

# === Scheduling ===
apscheduler>=3.10.0

# === API Server (future) ===
fastapi>=0.109.0
uvicorn>=0.27.0
websockets>=12.0

# === Testing ===
pytest>=8.0.0
pytest-asyncio>=0.23.0
pytest-cov>=4.1.0

# === Development ===
black>=24.0.0
isort>=5.13.0
mypy>=1.8.0
